backend/requirements.txt
Flask==3.1.0
python-dotenv==1.1.0
requests==2.32.3
requests-file==2.1.0
Werkzeug==3.1.3
